refactor migrate, seed

+++ database/migrations/2021_03_10_072301_create_shippings_table.php
@@ -24,7 +24,7 @@ class CreateShippingsTable extends Migration
             $table->string('mobile',100)->nullable();
             $table->string('creator',100)->nullable();
             $table->string('slug',100)->nullable();
-            $table->integer('status')->default(1);
+            $table->tinyInteger('status')->default(1);
             $table->timestamps();
         });
     }
